# gh-119396: Optimize PyUnicode_FromFormat() UTF-8 decoder