@vstinner 2024-05-22
cc @serhiy-storchaka