@vstinner 2024-05-22
I enabled automerge. Thanks for the review @serhiy-storchaka.